requirements.txt
uvicorn
fastapi
pymupdf
python-multipart
langchain
faiss-cpu
sentence-transformers
langchain-community
google-generativeai
python-dotenv
pinecone-client
youtube-transcript-api
qdrant-client